IoTMedicamentos/requirements.txt
python-dotenv>=1.0.0
pandas>=2.0.0
